Remove Project References
There were places in the build that used collections of projects like
coreModules that are no longer supported. There are also project
references that use a different name. For example rather than using ':' it
uses '-'.

This commit removes those references so that the build will pass, but
temporarily removes some functionality.